LayoutTests/accessibility/aria-scrollbar-role.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div tabindex=0 id="scroller" aria-valuenow="55" role="scrollbar" aria-orientation="vertical">scrollbar</div>


<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the ARIA scrollbar role works correctly");

    if (window.accessibilityController) {

        document.getElementById("scroller").focus();
        var scroller = accessibilityController.focusedElement;
        shouldBe("scroller.role", "'AXRole: AXScrollBar'");
        shouldBe("scroller.intValue", "55");
        shouldBe("scroller.orientation", "'AXOrientation: AXVerticalOrientation'");

        document.getElementById("scroller").setAttribute("aria-orientation", "horizontal");
        shouldBe("scroller.orientation", "'AXOrientation: AXHorizontalOrientation'");
    }

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
